feat(evaluations): dynamic hierarchical evaluation structure for primary apc report cards
Le Directeur configure lui-meme la grille d'evaluation d'un niveau, et le
bulletin PDF s'y adapte : plus aucun tableau de matieres code en dur.

Modele (migration AddHierarchicalEvaluationStructure, strictement additive)
- Subject.ParentSubjectId : un domaine porte des activites. DEUX niveaux au
  plus - le bulletin n'imprime qu'une colonne de regroupement, dont le RowSpan
  se calcule sur le nombre d'activites.
- Subject.MaxScore NULL veut dire "suis le bareme du cycle", pas "pas de
  bareme". Un defaut fixe a 20 aurait releve le plafond de tout le primaire,
  reste a /10. Resolution unique : GradeCalculator.EffectiveMaxScore.
- DisplayOrder, Column1Header, Column2Header.
- Index unique passe a (SchoolId, Level, ParentSubjectId, Name, IsDeleted) avec
  NULLS NOT DISTINCT : sans cette clause, ajouter une colonne nullable a la cle
  aurait fait cesser a l'index d'interdire deux matieres de meme nom au meme
  niveau. Les deux moities de la regle sont testees contre un vrai PostgreSQL.

Calcul
- Les lignes de baremes differents sont ramenees au bareme du bulletin avant
  d'etre moyennees (GradeCalculator.Rebase) : declarer "Competences /60" n'est
  pas declarer un poids, le poids reste le coefficient. L'appreciation se
  decide sur le pourcentage de reussite - une seule echelle pour tous les Sur.
- Le plafond de saisie suit la matiere : saisie unitaire, correction, import
  Excel, feuille de notes generee et ecran de notes.
- Un domaine ne se note pas : la note se saisit sur ses activites.

Bulletin
- EvaluationStructureBuilder imprime la grille ENTIERE, cases vides comprises,
  comme les modeles officiels. Un niveau sans hierarchie retombe sur les
  tableaux d'origine, inchanges.
- Rendu verifie sur les trois grilles reelles (CI-CP, CE1-CE2, CM1-CM2),
  chacune sur une seule page A5.

Ecran /matieres : onglet "Structure d'evaluation" - domaines, activites,
bareme par ligne, reorganisation, entetes des colonnes du bulletin.

Tests : 878 unitaires, 188 d'integration, 595 fonctionnels - tous verts.

+++ src/SamaEcole.Application/Common/Interfaces/IGradeSheetExcelGenerator.cs
@@ -10,7 +10,13 @@ namespace SamaEcole.Application.Common.Interfaces;
 /// </summary>
 public interface IGradeSheetExcelGenerator
 {
-    byte[] Generate(IReadOnlyList<GradeSheetStudentRow> rows, int gradingScale);
+    /// <summary>
+    /// <paramref name="gradingScale"/> est le barème de la MATIÈRE visée — celui que l'école lui a
+    /// fixé (grilles par compétences : /40, /60, /24…) ou, à défaut, celui du cycle de la classe. Il
+    /// gouverne à la fois l'en-tête des colonnes et la validation Excel des cellules, qui doivent dire
+    /// la même borne que celle appliquée à la réimportation.
+    /// </summary>
+    byte[] Generate(IReadOnlyList<GradeSheetStudentRow> rows, decimal gradingScale);
 }
 
 /// <summary>Une ligne de la feuille à générer : élève + ses notes déjà saisies (null si pas encore renseignée).</summary>
